Refactoring of Profiles.unit.test.ts - Refactored old Profiles.unit.test.ts file - Moved profiles tests out of extension.unit.test.ts file
