+++ __mocks__/mockCreators/datasets.ts
@@ -13,6 +13,7 @@ import { ZoweDatasetNode } from "../../src/dataset/ZoweDatasetNode";
 import * as vscode from "vscode";
 import * as imperative from "@zowe/imperative";
 import * as globals from "../../src/globals";
+import { removeNodeFromArray } from "./shared";
 
 export function createDatasetSessionNode(session: imperative.Session, profile: imperative.IProfileLoaded) {
     const datasetNode = new ZoweDatasetNode("sestest", vscode.TreeItemCollapsibleState.Expanded,
@@ -26,10 +27,12 @@ export function createDatasetTree(sessionNode: ZoweDatasetNode, treeView: any): 
     const testDatasetTree = {
         mSessionNodes: [sessionNode],
         mFavorites: [],
+        mRecall: [],
         treeView,
         addSession: jest.fn(),
         addHistory: jest.fn(),
         addRecall: jest.fn(),
+        addFavorite: jest.fn(),
         getHistory: jest.fn(),
         getRecall: jest.fn(),
         refresh: jest.fn(),
@@ -43,6 +46,7 @@ export function createDatasetTree(sessionNode: ZoweDatasetNode, treeView: any): 
         getTreeView: jest.fn().mockImplementation(() => treeView),
         searchInLoadedItems: jest.fn(),
         removeFavorite: jest.fn(),
+        deleteSession: jest.fn(),
         removeRecall: jest.fn(),
         enterPattern: jest.fn(),
         initializeFavorites: jest.fn(),
@@ -56,6 +60,12 @@ export function createDatasetTree(sessionNode: ZoweDatasetNode, treeView: any): 
         getSession: jest.fn(),
         getProfiles: jest.fn()
     };
+    testDatasetTree.addFavorite.mockImplementation((newFavorite) => testDatasetTree.mFavorites.push(newFavorite));
+    testDatasetTree.addRecall.mockImplementation((newRecall) => testDatasetTree.mRecall.push(newRecall));
+    testDatasetTree.removeRecall.mockImplementation((badRecall) => testDatasetTree.mRecall.splice(testDatasetTree.mRecall.indexOf(badRecall), 1));
+    testDatasetTree.getRecall.mockImplementation(() => { return testDatasetTree.mRecall });
+    testDatasetTree.deleteSession.mockImplementation((badSession) => removeNodeFromArray(badSession, testDatasetTree.mSessionNodes));
+    testDatasetTree.removeFavorite.mockImplementation((badFavorite) => removeNodeFromArray(badFavorite, testDatasetTree.mFavorites));
 
     return testDatasetTree;
 }
