+++ __mocks__/mockCreators/uss.ts
@@ -14,19 +14,22 @@ import * as imperative from "@zowe/imperative";
 import * as vscode from "vscode";
 import * as globals from "../../src/globals";
 import { ZoweTreeProvider } from "../../src/abstract/ZoweTreeProvider";
-import { USSTree } from "../../src/__mocks__/USSTree";
 import { getIconByNode } from "../../src/generators/icons";
+import { removeNodeFromArray } from "./shared";
 
 export function createUSSTree(favoriteNodes: ZoweUSSNode[], sessionNodes: ZoweUSSNode[], treeView?: vscode.TreeView<ZoweTreeProvider>): any {
     const newTree = {
         mSessionNodes: [...sessionNodes],
         mFavorites: favoriteNodes,
+        mRecall: [],
+        mSessions: [],
         addSession: jest.fn(),
         refresh: jest.fn(),
         removeRecall: jest.fn(),
         openItemFromPath: jest.fn(),
         addRecall: jest.fn(),
         getRecall: jest.fn(),
+        deleteSession: jest.fn(),
         checkCurrentProfile: jest.fn(),
         refreshElement: jest.fn(),
         getChildren: jest.fn(),
@@ -38,6 +41,13 @@ export function createUSSTree(favoriteNodes: ZoweUSSNode[], sessionNodes: ZoweUS
         setItem: jest.fn(),
         addHistory: jest.fn()
     }
+    newTree.addFavorite.mockImplementation((newFavorite) => newTree.mFavorites.push(newFavorite));
+    newTree.deleteSession.mockImplementation((badSession) => removeNodeFromArray(badSession, newTree.mSessionNodes));
+    newTree.removeFavorite.mockImplementation((badFavorite) => removeNodeFromArray(badFavorite, newTree.mFavorites));
+    newTree.addRecall.mockImplementation((newRecall) => newTree.mRecall.push(newRecall));
+    newTree.removeRecall.mockImplementation((badRecall) => newTree.mRecall.splice(newTree.mRecall.indexOf(badRecall), 1));
+    newTree.getRecall.mockImplementation(() => { return newTree.mRecall });
+
     return newTree;
 }
 
@@ -50,7 +60,7 @@ export function createUSSNode(session, profile) {
 }
 
 export function createUSSSessionNode(session: imperative.Session, profile: imperative.IProfileLoaded) {
-    const zoweUSSNode = new ZoweUSSNode("parent", vscode.TreeItemCollapsibleState.Collapsed, null,
+    const zoweUSSNode = new ZoweUSSNode("sestest", vscode.TreeItemCollapsibleState.Collapsed, null,
         session, "/", false, profile.name, undefined, profile);
     zoweUSSNode.fullPath = "test";
     zoweUSSNode.contextValue = globals.USS_SESSION_CONTEXT;
@@ -63,7 +73,7 @@ export function createUSSSessionNode(session: imperative.Session, profile: imper
 }
 
 export function createFavoriteUSSNode(session, profile) {
-    const ussNodeF = new ZoweUSSNode("[profile]: usstest", vscode.TreeItemCollapsibleState.Expanded, null, session, null, false, profile.name);
+    const ussNodeF = new ZoweUSSNode("[sestest]: usstest", vscode.TreeItemCollapsibleState.Expanded, null, session, null, false, profile.name);
     const mParent = new ZoweUSSNode("Favorites", vscode.TreeItemCollapsibleState.Expanded, null, session, null, false, profile.name);
     mParent.contextValue = globals.FAVORITE_CONTEXT;
     ussNodeF.contextValue = globals.DS_TEXT_FILE_CONTEXT + globals.FAV_SUFFIX;
