+++ i18n/sample/src/Profiles.i18n.json
@@ -3,6 +3,14 @@
     "loadNamedProfile.error.profileName": "Could not find profile named: ",
     "loadNamedProfile.error.period": ".",
     "createNewConnection.invalidzosURL": "Please enter a valid host URL in the format 'company.com'.",
+    "addSession.quickPickOption": "Choose \"Create new...\" to define a new profile or select an existing profile to Add to the USS Explorer",
+    "enterPattern.pattern": "No selection made.",
+    "createNewConnection.option.prompt.profileName.placeholder": "Connection Name",
+    "createNewConnection.option.prompt.profileName": "Enter a name for the connection",
+    "createNewConnection.enterprofileName": "Profile Name was not supplied. Operation Cancelled",
+    "addSession.log.debug.createNewProfile": "User created a new profile",
+    "createZoweSession.log.debug.selectProfile": "User selected profile ",
+    "createZoweSession.log.debug.cancelledSelection": "User cancelled profile selection",
     "createNewConnection.zosmfURL": "No valid value for z/OS URL. Operation Cancelled",
     "createNewConnection.undefined.port": "Invalid Port number provided or operation was cancelled",
     "createNewConnection.undefined.username": "Operation Cancelled",
